test: move tests in right package
