--- operate/importer-8_4/src/main/test/java/io/camunda/operate/zeebeimport/processors/fni/FNITransformerCacheTreePathTest.java
@@ -1,88 +1,0 @@
-/*
- * Copyright Camunda Services GmbH
- *
- * BY INSTALLING, DOWNLOADING, ACCESSING, USING, OR DISTRIBUTING THE SOFTWARE (“USE”), YOU INDICATE YOUR ACCEPTANCE TO AND ARE ENTERING INTO A CONTRACT WITH, THE LICENSOR ON THE TERMS SET OUT IN THIS AGREEMENT. IF YOU DO NOT AGREE TO THESE TERMS, YOU MUST NOT USE THE SOFTWARE. IF YOU ARE RECEIVING THE SOFTWARE ON BEHALF OF A LEGAL ENTITY, YOU REPRESENT AND WARRANT THAT YOU HAVE THE ACTUAL AUTHORITY TO AGREE TO THE TERMS AND CONDITIONS OF THIS AGREEMENT ON BEHALF OF SUCH ENTITY.
- * “Licensee” means you, an individual, or the entity on whose behalf you receive the Software.
- *
- * Permission is hereby granted, free of charge, to the Licensee obtaining a copy of this Software and associated documentation files to deal in the Software without restriction, including without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is furnished to do so, subject in each case to the following conditions:
- * Condition 1: If the Licensee distributes the Software or any derivative works of the Software, the Licensee must attach this Agreement.
- * Condition 2: Without limiting other conditions in this Agreement, the grant of rights is solely for non-production use as defined below.
- * "Non-production use" means any use of the Software that is not directly related to creating products, services, or systems that generate revenue or other direct or indirect economic benefits.  Examples of permitted non-production use include personal use, educational use, research, and development. Examples of prohibited production use include, without limitation, use for commercial, for-profit, or publicly accessible systems or use for commercial or revenue-generating purposes.
- *
- * If the Licensee is in breach of the Conditions, this Agreement, including the rights granted under it, will automatically terminate with immediate effect.
- *
- * SUBJECT AS SET OUT BELOW, THE SOFTWARE IS PROVIDED “AS IS”, WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE, AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
- * NOTHING IN THIS AGREEMENT EXCLUDES OR RESTRICTS A PARTY’S LIABILITY FOR (A) DEATH OR PERSONAL INJURY CAUSED BY THAT PARTY’S NEGLIGENCE, (B) FRAUD, OR (C) ANY OTHER LIABILITY TO THE EXTENT THAT IT CANNOT BE LAWFULLY EXCLUDED OR RESTRICTED.
- */
-package io.camunda.operate.zeebeimport.processors.fni;
-
-import static io.camunda.operate.zeebeimport.processors.fni.FNITransformerTest.createZeebeRecord;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.eq;
-import static org.mockito.Mockito.times;
-import static org.mockito.Mockito.when;
-
-import io.camunda.operate.zeebeimport.cache.FNITreePathCacheCompositeKey;
-import io.camunda.operate.zeebeimport.cache.TreePathCache;
-import io.camunda.zeebe.protocol.record.intent.ProcessInstanceIntent;
-import io.camunda.zeebe.protocol.record.value.BpmnElementType;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.EnumSource;
-import org.junit.jupiter.params.provider.EnumSource.Mode;
-import org.mockito.Mockito;
-
-public class FNITransformerCacheTreePathTest {
-
-  private FNITransformer fniTransformer;
-  private TreePathCache mockTreePathCache;
-
-  @BeforeEach
-  public void setup() {
-    mockTreePathCache = Mockito.mock(TreePathCache.class);
-    when(mockTreePathCache.resolveParentTreePath(any()))
-        .thenAnswer(
-            invocationOnMock -> {
-              final FNITreePathCacheCompositeKey compositeKey = invocationOnMock.getArgument(0);
-              return String.format(
-                  "%d/%d", compositeKey.processInstanceKey(), compositeKey.flowScopeKey());
-            });
-    fniTransformer = new FNITransformer(mockTreePathCache);
-  }
-
-  @ParameterizedTest
-  @EnumSource(
-      value = BpmnElementType.class,
-      names = {"SUB_PROCESS", "PROCESS", "EVENT_SUB_PROCESS", "MULTI_INSTANCE_BODY"},
-      mode = Mode.INCLUDE)
-  void shouldCacheContainerFNI(final BpmnElementType elementType) {
-    // given
-    final var time = System.currentTimeMillis();
-    final var record =
-        createZeebeRecord(time, ProcessInstanceIntent.ELEMENT_ACTIVATING, elementType);
-
-    // when
-    fniTransformer.toFlowNodeInstanceEntity(record, null);
-
-    // then
-    Mockito.verify(mockTreePathCache, times(1)).cacheTreePath(any(), eq("1/3/4"));
-  }
-
-  @ParameterizedTest
-  @EnumSource(
-      value = BpmnElementType.class,
-      names = {"SUB_PROCESS", "PROCESS", "EVENT_SUB_PROCESS", "MULTI_INSTANCE_BODY"},
-      mode = Mode.EXCLUDE)
-  void shouldNotCacheNonContainerFNI(final BpmnElementType elementType) {
-    // given
-    final var time = System.currentTimeMillis();
-    final var record =
-        createZeebeRecord(time, ProcessInstanceIntent.ELEMENT_ACTIVATING, elementType);
-
-    // when
-    fniTransformer.toFlowNodeInstanceEntity(record, null);
-
-    // then
-    Mockito.verify(mockTreePathCache, times(0)).cacheTreePath(any(), any());
-  }
-}
